pyproject.toml
[project]
name = "lineup-generator"
version = "0.1.0"
description = "Generate lineup guide PNGs for LED screens built from LED tiles"
requires-python = ">=3.10"
dependencies = [
  "pillow>=10.0.0",
  "streamlit>=1.30.0",
  "pydantic>=2.5.0",
]

[tool.ruff]
line-length = 100
